requirements.txt
numpy==1.26.4
joblib==1.3.2
streamlit==1.31.1
scikit-learn==1.4.1.post1
joblib==1.3.2
Pillow==9.4.0
